Final version 1.0

+++ app/Http/Controllers/ProfileController.php
@@ -8,6 +8,14 @@ use Illuminate\Validation\Rule;
 
 class ProfileController extends Controller
 {
+    public function index()
+    {
+
+        return view('profile.index',[
+            'users' => User::paginate(15)
+        ]);
+    }
+
     public function show(User $user)
     {
 
@@ -19,6 +27,11 @@ class ProfileController extends Controller
 
     public function edit(User $user)
     {
+        if(auth()->user()->id !== $user->id && auth()->user()->cannot('admin'))
+        {
+            abort('403');
+        }
+
         return view('profile.edit', [
             'user' => $user
         ]);
@@ -26,6 +39,11 @@ class ProfileController extends Controller
 
     public function update(User $user)
     {
+        if(auth()->user()->id !== $user->id && auth()->user()->cannot('admin'))
+        {
+            abort('403');
+        }
+
         $attributes = request()->validate([
             'username' => ['required', 'max:255', Rule::unique('users','username')->ignore($user->id)],
             'phone_nr' => 'required|regex:/^([0-9\s\-\+\(\)]*)$/|min:8',
@@ -39,9 +57,26 @@ class ProfileController extends Controller
 
     public function destroy(User $user)
     {
+        if(auth()->user()->id !== $user->id && auth()->user()->cannot('admin'))
+        {
+            abort('403');
+        }
+
         auth()->logout();
         $user->delete();
 
         return redirect('/')->with('success', 'Profils izdzēsts!');
+    }
+
+    public function destroyAdmin(User $user)
+    {
+        if(auth()->user()->id !== $user->id && auth()->user()->cannot('admin'))
+        {
+            abort('403');
+        }
+
+        $user->delete();
+
+        return back()->with('success', 'Lietotājs izdzēsts!');
     }
 }
